moved script generation into utils
