+++ src/App.tsx
@@ -7,10 +7,9 @@ import {
 } from './lib/TriggerDataManager.types';
 import { readJsScript } from './lib/io/read-js-script';
 import { readJsonScript } from './lib/io/read-json-script';
-import { writeJsScript } from './lib/io/write-js-script';
 import { getSetting, TEXT_SIZES } from './lib/settings-storage';
 import { FONT_SIZE_SETTING, SETTINGS_KEY, VIEWPORT_SETTING } from './lib/settings-storage.types';
-import { validateTimes, formatSkins } from './lib/util';
+import { validateTimes, formatSkins, generateScript } from './lib/util';
 import { CONSTRAINT } from './lib/constraints';
 import { GLOBAL_STYLES } from './styles';
 
@@ -301,7 +300,7 @@ export class App extends React.Component {
         throw new Error('Triggers contain invalid times!');
       }
 
-      const script = writeJsScript(activeTab, this.triggerManager.data as TriggerDataLookup);
+      const script = generateScript(activeTab, this.triggerManager.data as TriggerDataLookup);
       return await navigator.clipboard.writeText(script);
     } catch (error) {
       if (error instanceof Error) {
